Chore: constant폴더에서 HtmlSelector.js 모으기
