+++ src/constant/constant.js
@@ -1,8 +1,1 @@
 export const MAX_LOCAL_STORAGE = 10;
-export const RECENT_SEARCH_LIST = 'recentSearchList';
-
-export const SEARCH_BOX = {
-  RECENT__DELETE: 'recent__delete',
-  RECENT__TERM: 'recent__term',
-  RECENT__DELETE__ALL: 'recent__delete-all',
-};
