+++ src/constant/htmlSelector.js
@@ -2,3 +2,10 @@ export const POP_UP = {
   show: 'pop-up-show',
   hidden: 'pop-up-hidden',
 };
+export const RECENT_SEARCH_LIST = 'recentSearchList';
+
+export const SEARCH_BOX = {
+  RECENT__DELETE: 'recent__delete',
+  RECENT__TERM: 'recent__term',
+  RECENT__DELETE__ALL: 'recent__delete-all',
+};
